Link previews (prototype 1 + feature flags)
Display a "preview" version of a target article in a native component.

* 60/40 proportion for page text extract vs. image
* Tapping anywhere within the card (including title link) takes you to the page
* Added rudimentary gesture detector:
** Fling up: go to page
** Fling down: dismiss
* Improved splitting of sentences by using client-side BreakIterator.
* Displaying a single sentence by default, and two sentences if the first
sentence is shorter than four lines.
* Added three-way feature flag:
** 50% of users (and the production app) will see the original behavior (no preview)
** 25% of users will see prototype 1 (this patch)
** 25% of users will see prototype 2 (the next patch)
* Refactored feature-flagging code a bit.

Note: In the first few revisions of this patch, I added support for "long-press"
handlers to our ObservableWebView, but ended up not using it in the final version.
However, I left it in there anyway, since it will surely come in handy later.

Bug: T95030 T96068
Change-Id: I599c9948184e478dc8b377499680621af3e1caf0