Hygiene: reorder XML for views so that the android:id is on the first line
I needed something menial to do to relax, so here it is. I reordered a ton of
the XML so that the android:id is almost always on top, and separated out some
attributes so that they're on multiple lines. There should be no user-facing
changes as a result of this, but it lets new developers easier find the ID of
a view at a glance.

Change-Id: Idb61079339be45e366050c2ffa53fd12fd1fab8e