Change version number format to include versionCode as subminor number
I'm proposing to include the versionCode into the version name format.
This will make version names more distinct, albeit for now there's a bit of redundancy.

Example:
2.0-beta-2015-04-09 becomes
2.0.100-beta-2015-04-09

I think eventually we could even consider getting rid of the date portion but
I'll keep it for now to make the transition smoother.

Had to also change the make-release.py script since it generated the version
name itself based on the date the command was run.
I'd like to get away from that and use the apk file to derive the version name.
It uses the versionCode from the build.gradle file.

To be able to extract the versionName from the apk it needs to know the location
of the aapt tool, which is under $ANDROID_HOME/build-tools/<build_tools_version>.
Again, the build_tools_version is retrieved from the build.gradle file.

Bug: T95584
Change-Id: Icb6a076181bb1e696be5b646a6891c442d4d6d69