Slightly decrease minimum text size on Share a Fact cards.
This patch slightly decreases the minimum text size on Share a Fact cards.
This allows medium-length quotes to be fit on the cards.

Also increased the top padding to be similar to the left/right padding per Vibha's request.
It's not exactly the same value because there is some ascender height space above the cap height.

Bug: T96398
Change-Id: Ia090a6a4de50c653c3abc189a00ba0c30bb310b1