Run disambiguation links through URLDecoder
This patch changes the app to run disambiguation links through a URL decoder,
and falls back to using the old system of displaying the encoded URL if
there's some problem with the decoding.

In some cases URL encoded links were shown to the user, which in the best
case just looks bad and is confusing, but in the worst case actually causes
the links to be incorrect and navigation to fail. This patch fixes that.

Bug: T97200
Change-Id: I4ed27184ad5ebb5bfe8ddfbea4b2fbc706c25918