Handle some more login API errors.
This patch handles two additional errors in the login workflow, "Illegal" and
"WrongPluginPass". The API returns "Illegal" if the username the user tried
to log in with is invalid, and returns "WrongPluginPass" if an authentication
extension like CentralAuth blocked the login. Our API experts advised us that
API consumers should treat WrongPluginPass and WrongPass as the same, so this
patch does that.

Bug: T97477
Change-Id: Ib5a45fb6f02a4d0889455f45bb713f4213443aa8