Make some PersistenceHelpers set the namespace of pages as appropriate
Currently, SavedPagePersistenceHelper and HistoryEntryPersistenceHelper do
not set the namespace of a page appropriate, setting it to null. I'm not sure
why it was implemented that way.

This patch makes SavedPagePersistenceHelper and HistoryEntryPersistenceHelper
set the namespaces of pages appropriately.

Fortunately, the method in which the database is upgraded ensures the old
entries are still compatible with the new system. They'll continue to suffer
the same problem where the pages are constructed with null namespaces, but
otherwise they'll work as intended without any tweaking.

Bug: T96944
Change-Id: I64b6ae2d72caa9c535f6a2a5627dc7a23eb2f6f7