Fix TextView directionality issue for mixed LTR and RTL strings
Some TextViews are displaying mixed LTR and RTL text incorrectly, formatting
it as if it's an LTR string, which results in incorrect display. This patch
fixes those TextViews to display as appropriate. Note: the fix was confirmed
to be the correct way of displying mixed text by a Hebrew speaker (Matanya
on the Phabricator task.

Bug: T97958
Change-Id: Idec520cbc9ee8db8d691c2f487d2bfd0042f7bae