Fix sharing of images to Facebook app.
The Facebook app requires us to explicitly specify a mimetype of
image/jpeg when sharing an image, whereas we were previously specifying
"*/*".

I simply modified the shareImage() function to always specify
"image/jpeg", since that's all it can ever be.

Note that the Facebook app still doesn't seem to do anything with the
EXTRA_TEXT that we pass into the share intent, but we can't do much about
it, since it's up to the receiving app to handle the text correctly.

Bug: T97858
Change-Id: Ie9d333d7105c3acfa7004df86a8a17dbd6e78c1e