Switch to System.nanoTime() for time measurements in SearchResultsFragment
Presently the app uses System.currentSystemMillis() to measure time
intervals. This method is subject to changes in the system clock, which can
throw off time measurements rather substantially. This patch changes
SearchResultsFragment to use System.nanoTime() instead, which is not subject
to changes in the system clock and actually has a higher level of precision
anyway. This should fix the issue where some search result sets were
allegedly being displayed many days before the user even typed the query in.

Bug: T99577
Change-Id: I843e63a0d96edcf6534c04c97cdaf5c4348bedc2