Fix toolbar translucency when TOC closed
SimpleDrawerListener.onDrawerClosed is not called unless the DrawerLayout is
closed after reaching a fully open state (i.e., not when slightly opened and
closed again).

This patch fixes the toolbar translucency checks so that setForceNoFace is
reset to false whenever the drawer is out of sight rather than relying
on reaching onDrawerClosed.

Bug: T100382
Change-Id: Ia3507fb9391d64c553801cd46c06b383c88f8bba