Tweak local cache policy in app.
Presently, we load the cached version of a page regardless of how the page
was reached. This is a bit too aggressive, since some pages are expected
by users to be updated relatively frequently, and users don't want to use
(or aren't aware of) our pull-to-refresh feature.

This patch makes it so that pages are loaded from cache only when
navigating back through the current backstack of pages, or when coming
back from savedInstanceState. For "forward" navigation of pages, e.g.
clicking links, clicking history items, etc. will no longer use cache for
now.

This will be followed by a more comprehensive patch where we request the
current revision of the page from the network, and load the full page from
the network only if the revision is newer than what's cached.

Bug: T101845
Change-Id: Ide6c35cdedfa8c6ffc2ac3f4d89332270ecb0ef2