Fix issue where app goes to main page on device rotation
When a config change occurred it saved a null value for the "language" in onSaveInstanceState.
Subsequently, when the activity was recreated, it used the null value and compared it with
app.getAppOrSystemLanguageCode(), which is always set to a non-null string (in my case "en").

Why is this affecting the page we're on? Because we use the language value to determine if
the language has changed and if we need to reset the site, which also goes to the sites main page.

Thus, it looks like we have to go back to Start, and don't collect our $400.

Bug: T102097
Change-Id: Ib2a4b8fcd66f601017ff486484ec3bcfdd5a27af