Simplify and fix inconsistencies in share-a-fact handlers.
This patch unifies the TextSelected and NoTextSelected handlers, since
both are now using the same call to the Bridge to retrieve selected text
inside the WebView.

In the case of no text highlighted, we switch to fetching the actual
content of the first <p> element from the WebView, instead of extracting
the HTML in Java. This also makes it consistent with iOS.

Bug: T99151 T99144 T96493 T98994
Change-Id: I5d5be3b025643e004402067099abe6d00621ce0c