Make sure random feature flag ID is positive.
We currently do a/b testing by taking the feature flag ID modulo a certain
number, and using the result to switch between features.

However, it's not immediately obvious that the feature flag ID can be a
*negative* number, and might produce a negative value after the modulo
operation.  This patch ensures that the feature flag ID will always be a
positive number.

Change-Id: I43f0af1e22b5f9a49db77ee6151821852021fb77