Fix erroneous description in articles without a description.
This was happening because we're performing asynchronous animations on the
Page Title text view, and then fading in the Description when the Title
animation finishes. If we start loading a new page in the middle of the
title animation, the fade-in of the description will still happen, even if
the new page doesn't have a description.

This patch fixes the issue by explicitly canceling the animation whenever
a new page is loaded.

Bug: T100173
Change-Id: I9c2ece0a8292be080857ce5fc1864b8d04616f16