Add language selection button to search bar
Currently, switching the app language requires accessing the preferences screen
via the nav drawer and selecting the desired language from there.  That's
inconvenient when one is in the middle of something, a search for instance.

This patch puts a language selection button right on the search bar to make it
more convenient to change the app language on the fly.

Bug: T73136
Change-Id: I1b638ffe8909307aac9e4355a7fceca1631643c7