Add custom View for displaying errors consistently.
Added a ThrowableUtil class that provides functions to get well-formed
error messages for any Throwable. These messages may be used in Toasts,
SnackBars, custom Views, etc.

Added a custom View that uses the above functions, and formats the
messages in a consistent way. This custom View may be used anywhere that
we want to display errors.

As an example, hooked up the new functionality into the Search fragment.
In subsequent patches, it should be used in additional places, too.

Note: this is dependent on the corresponding patch in java-mwapi:
https://gerrit.wikimedia.org/r/221813

Bug: T102464
Change-Id: Ie53a0c40309b043f8f59377032cbeb6d88e5e6ee