Fix full-text search in main Search bar.
The issue was that the addition of "morelike" searching didn't
differentiate between searching from the Search bar, and searching for
Read More suggestions. The "morelike" search, if enabled, should only
apply to Read More suggestions.

Therefore, this patch adds a parameter to the full-text search task to
specify whether to perform a "morelike" search, which the Search bar now
explicitly sets to false.

Bug: T104293
Change-Id: Ibfb795c575dead30d86e720ab91d7a1056a2bb57