Avoid NPE in JsonPageLoadStrategy.onSaveInstanceState
This cannot be easily reproduced but according to the stack trace this happens
when backStack == null during onSaveInstanceState.
This patch set the back stack to an empty list right in the JsonPageLoadStrategy constructor. While it seems a bit wasteful since it should get set later anyways it's probably the safer thing to do.

Bug: T104090
Change-Id: Ib359cce9a508f3d9640af5ae23c38aecf923d17a