Smooth gradients in Lead Image and Gallery.
When viewed on top of especially light images, the gradients that we
display on top of the Lead Image, as well as the Gallery images, can have
a noticeably abrupt "end", because these gradients are linear.

This patch modifies the gradients to follow a cubic function, so that the
gradient trails off more smoothly than a linear function.

I've also removed the corresponding unused XML gradient drawables.

Bug: T104808
Change-Id: Ic66637ea649ea7945d9ffe688c5d03b20f47810d