Add length cutoff for text snippets sent over JS bridge
Currently, if a user selects a very large text block and attempts to share it,
the share action will fail and/or hang or crash the app.

Example: Visit article "Victor Rousseau Emanuel," select all text, and attempt
to share the selection.

This patch truncates the selection sent over the JS bridge at 250 characters
(which appears to be approximately the max that the share-a-fact card can display)
in order to eliminate out-of-memory issues.

Bug: T104527
Change-Id: I371a8df1261ca945b3ed32e9eaacaabb74ae43f5