Don't display "no results found" before running full text search
Currently, we display the results of a prefix search (including "no results found"
in the case of an empty result list) before falling back to a full text search.
If, as is frequently the case, the full text search yields results when the prefix
search didn't, this results in "no results found" flickering momentarily before
results are displayed.

This patch prevents this behavior by only calling displayResult() if the prefix
search results are non-empty before falling back to full text search.

Bug: T104647
Change-Id: Iaa4a5fbe2b7bdb16f32869175d6a4259f67b17df