Add Share A Fact onboarding properties to funnel
Report whether Share A Fact onboarding tips were enabled and which were
shown to the user. This reporting is used to measure the success of the
onboarding experience.

Rather than create a new schema, we leverage the existing
MobileWikiAppShareAFact schema with new properties for feature enabled,
and tips shown.

Logging verified on
deployment-eventlogging02.eqiad.wmflabs:/var/log/eventlogging/all-events.log.

Bug: T89989
Change-Id: I3d1d5d49c5619f8d6b1b4014079bab3e8336b5ac