Add developer settings screen
* Add secret developer settings UI.
* Add EditTextAutoSummarizePreference and IntPreference.
* Move ThemedActionBarActivity to activity subpackage.

A developer settings screen is endlessly useful for changing otherwise
internal state on the fly without the use of a debugger or coding. It
also encourages dev culture.

This patch exposes the preference state for various shared preferences
used to A/B test layouts but further enhancements may yet be added.

Change-Id: I914a7e6d8600177860e252a8490b3d6ea6691216