Hygiene: Refactor settings code
Remove a lot of duplicated code between SettingsFragment and SettingsActivityGB.

This patch adds an abstraction layer over PreferenceActivity and PreferenceFragment,
called PreferenceHostCompat. This is unfortunately needed since, as of this writing,
there is no support-v4 variant of PreferenceFragment, or common base class for
preference related functionality provided by the Android support libraries.

Change-Id: If73081533326bcd5202b0e80a18567f7030c7201