Add article URL to share-as-text
Share-a-fact used to share the article URL when no text was selected; this
is no longer the case since we default to sharing the first paragraph.
Since users commonly obtained the article URL this way, we need to put it
back.

This patch does that.

Additionally, it removes a redundant snippet length check.

Bug: T103951, T104704
Change-Id: I287c7270973ddf0e1fd752fbcef1f18c0a6ba745