Improve sequencing of ToC and Text-highlight tooltips.
Previously, if the ToC tutorial was shown to the user, then the Highlight
tutorial wouldn't be shown until the app is destroyed and relaunched,
which is not a predictable interval.

This patch guarantees that the Highlight tutorial is shown after the ToC
tutorial, with a minimum time interval of 1 minute between the two
tutorials (so that the user isn't inundated).

Change-Id: I621d179f47c0b606ab24facec06a654e70373c83