Use TextInputLayout in Login and CreateAccount activities.
Wrapped EditText's in the Login and CreateAccount activities in
TextInputLayout, as well as the EditText in the Captcha entry view, and
the edit preview summary.

Also made PlainPasteEditText inherit from AppCompatEditText, to make sure
that any styling that we apply to EditText in the future will be applied
properly by AppCompat.

Note: did *not* modify the EditText in the History and Saved Pages
fragments, since they're being affected by a known bug in AppCompat:
https://code.google.com/p/android/issues/detail?id=175228
Once the above issue is fixed, we can finish those updates.

Bug: T102549
Change-Id: I72b1acd858c47ecb20dfc0b7b269c11aad5a99a9