Fix share-a-fact behavior where lead image license has no license info
Currently, in the event license info is unavailable for the lead image, the
share-a-fact card will default to showing the fair use icon.  This isn't the
desired behavior.

After this patch, share-a-fact will treat a situation in which no license info is
available the same as when no lead image exists: the selected (or default) text will
appear against a black background, and a creative commons license covering the card
will be indicated.

Bug: T105726
Change-Id: I65760f1d6cd42cafeec4a4d706d26ba79150afbf