Fix width of share-a-fact dialog contents.
Here's the problem:
The width of the "share as..." buttons is dictated by the width of the
preview image. The issue is that the preview image has a hard-coded width
in absolute pixels (not dp). Therefore, the image might appear larger on a
hdpi screen, or smaller on a xxhdpi screen.  When it appears smaller, it
causes the buttons to shrink, and thereby truncate the text of the
buttons, giving the user the impression that the buttons are identical.

This patch fixes the width and height of the ImageView used for displaying
the preview bitmap, so that it's displayed consistently on all display
densities, and makes the buttons sufficiently wide, as well.

Bug: T99966
Change-Id: Ibdecbfb8bccc31108c9fb8f3879ce6fb59985dca