Replace Toast and Crouton with Snackbar.
Unified the display of errors and warnings using Snackbar in place of
Toast and/or Crouton, while utilizing the new methods for converting
Throwables to plain strings.

Turned the Wikipedia Zero banner from a Crouton into a Snackbar. Was able
to keep the exact same formatting as the Crouton implementation, except
the banner now pops up from the bottom, instead of the top. This is the
only change that the Partnership team will need to approve.

Note: this will need a LOT of testing.

Bug: T102548

Change-Id: I4314a8d219a90415636cfaca867152b9251e551d