Add long-press menu to Search, Nearby, Disambig, and Read More.
Generalized the PageLongPressHandler to be usable by a ListView in
addition to a generic View. In the case of a ListView, we automatically
give it an OnItemLongClickListener that handles the long-press event.

Because the standard ListView doesn't provide the x-y coordinates of the
long-press event, I created WikiListView where it records the last known
x-y position from monitoring touch events.

Made BottomDialog inherit from AppCompatDialog, and turned some of its
initialization code into styles.

Did NOT add the long-press menu to History and Saved Pages, since they
already have a long-press action (to select multiple items for deletion).
We'll need to think about that part a bit further (future patch).

Bug: T103957
Change-Id: Ibbf613bfb50435f643787430da590621181f1d1b