Add neutral background splash screen
Dark mode users do not want to be blinded during app launch by a light
themed Activity. Unfortunately, Activity.setTheme() and
getWindow().setBackgroundDrawable() don't take effect soon enough. I can
find no other mechanism for changing the Activity theme dynamically.
This is a workaround patch that changes the launch screen theme to be a
neutral grey with a big W in it. The W keeps things interesting and
makes a promise to the user that good things come to those who wait.

Bug: T106392
Change-Id: I1539405fcf2ab097516e13ef38a6301bec58b77e