Adjust sampling rate for new funnels, in preparation for production.
Did a bit of refactoring of the sample rate logic. Funnels may now
pass their sampling rate directly into the super constructor. The base
Funnel class now does the job of fetching the sample rate from remote
config parameters. The name of the parameter for each funnel is now
derived from the actual name of the schema that the funnel uses, plus the
string "_rate". For example, the remote parameter for the SearchFunnel
would be "MobileWikiAppSearch_rate".

Since our current remote variables for the funnels either correspond
exactly to the hard-coded value, or don't exist, it's safe to rename them
in our code, and add them to the remote config file as needed.

- Added sample rate of 100 to GalleryFunnel.
- Added sample rate of 100 to AppLanguageSelectFunnel.
- Since LinkPreviewFunnel is still in Beta, gave it a sample rate of 10,
  since the number of these events from the previous Beta is a bit much
  (it wasn't being sampled at all). When this is ready for production,
  will increase it to 100, or even 1000.
- Added sample rate of 100 to SuggestedPagesFunnel.
- Added sample rate of 100 to TabFunnel.
- Added sample rate of 100 to DailyStatsFunnel.

Change-Id: I9b20f161abad228cb551a4e6805207907fbbc06c