Hygiene: disable incremental build optimize on CI
It is recommended[0] to disable pre-dexing when performing clean builds.
This patch provides a property hook for controlling pre-dex which when
unspecified defaults to disabled when executed from CI and true
otherwise.

An informal analysis was performed with this patch showing a nearly 2x
performance increase for clean builds:

1. <code>time for i in {1..10}; do time JENKINS_HOME=/home/stephen/tmp ./gradlew -q clean assembleDevDebug; done</code>
   4m38s total; 28s average iteration.
2. <code>time for i in {1..10}; do time ./gradlew -q clean assembleDevDebug; done</code>
   2m23s total; 14s average iteration.

Actual benefits when executed on the server may be more modest.

[0] Pro tip #4 from https://plus.google.com/+AndroidDevelopers/posts/ECrb9VQW9XP

Change-Id: I50f757ac2378c029fb47577b03c5dae49a0517c5