Nail down tab behavior for pages coming from intents.
Will now open pages sent from intents in a new tab.

Modified Back behavior so that, when the backstack in the current tab is
empty, the activity finishes (instead of animating into the tab list).
This is consistent with other browser apps, and allows correct back
behavior for links coming from an intent (i.e. a single Back press should
take the user back to the previous activity, not the tab list).

Made PageActivity be singleTask. According to the documentation, this is
"not appropriate for most applications"... but we are not most
applications. The rationale for this is that, since we now store our tab
and backstack state in SharedPrefs, it no longer makes sense to allow
multiple instances of PageActivity, since one instance that modifies the
tab state would make it out of sync with the others. And in any case, a
single instance of PageActivity will benefit the user by providing a
single point of access to the user's "state of research" in the app.

Bug: T107487
Change-Id: I242748438b73c25fccd99e5bad64db33c697d48e