Show "Display password" check box on login screen
If you make a mistake when typing in your password, it's nice to have a "Show
password" check box so you don't have to retype your entire password (which
would be quite the inconvenience on a mobile device if you have a long
password). This patch implements that, and it was super simple to do due to a
provided convenience method in Utils! :-)

Bug: T99680
Change-Id: I40d1e7e91a641308f2066408742288373c5f194e