Use standard ContextMenu for long-press actions.
This patch standardizes the long-press action for our WebView and
ListViews to use a standard ContextMenu (which pops up in the center of
the screen), instead of a custom PopupMenu. This makes it consistent with
other apps that implement long-pressing within a WebView, as well as
significantly simplifies the overall logic. As a bonus, this also makes
long-pressing on links work correctly in ICS.

Bug: T107839
Change-Id: I7a86dd6d1fc934cd4d9f88dede5bfd83d9b1bb5e