Halve RGB values of background colors in tables in night mode
Most colors are inverted in night mode.  However, the background colors
of some elements in tables  look strange when given this treatment, so
it was earlier decided to simply strip them as a temporary fix.  This
resulted, however, in a loss of useful information (see, e.g., the diagram
in the quick facts box for 'National Assembly of Pakistan').

This patch halves the RGB values of these elements rather than simply
stripping the colors in order to preserve necessary information.

Bug: T104101
Change-Id: I567089abf4d39ed1db562d51bd419f26d97d429d