Check if page is saved in DB when preparing 'save page' options menu item
Currently, the app doesn't reliably know when it's on a saved page. To fix
this, I removed the existing code dealing with page save state[1] and
implemented a check over the saved pages in the database when the options
menu is prepared to determine whether the current page is in fact a saved
page.

[1] This previous code did not actually keep track of page save state but
set it as saved when navigating from the saved pages fragment or when
refreshing after saving a page. As such, it failed to detect when the user
navigated to a saved page from other sources, e.g., an ordinary search.

Bug: T102293
Change-Id: I95ae9c11f1b29b16320359c8f0f0de52ecb3abc0