Improve marshalling of PageProperties when saving state.
One of the latest crash reports relating to Gson unmarshalling has yielded
something revealing: the PageProperties object contains a private instance
of a SimpleDateFormat object which, when marshalled, takes up a *lot* of
unnecessary space, and may possibly be contributing to errors when it's
unmarshalled.

This patch removes the SimpleDateFormat object from PageProperties, and
moves it to WikipediaApp (into a method that returns a new instance of
SimpleDateFormat for every call, since SimpleDateFormat is not
thread-safe), making it accessible to any PageProperties instance, and
helps to optimize the marshalling of the app's tab/backstack state.

Bug: T108617
Change-Id: I9c16b554cbfa84213abd892232ac086a1f469e98