Apply correct Material background to some buttons.
There's a predefined background attribute that already happens to exist in
the library, which we can apply to our own buttons, which will make their
background behave just like the system buttons in the Toolbar.

I went ahead and applied this background to the Language selection button,
as well as the Clear History button, and the secondary action in Nearby
items. There may be additional candidates to receive this background, but
this should do for now.

Bug: T108381
Change-Id: I0e4366b0fba115e0ea5751797c4c9120a3035658