Update SQL query generation to handle mainspace and other null params
This patch updates the generation of SQL query parameters to be passed
to the content provider by generating primaryKeySelection and
PrimaryKeySelectionArgs values based on the properties of a specific
object, handling null values for namespace (or any other field)
appropriately.

This eliminates the problems of duplicate history and saved page entries,
and allows deleting history and saved pages entries one at a time whether
the underlying pages are in mainspace (i.e., have a NULL value stored as
their namespace) or otherwise.

Bug: T102293 T105471 T105472 T107842
Change-Id: Ie715baefae7b76fb488e0910be5fcf4ad4593b4c