Maintain PageActivity instance across screen rotation.
Added android:configChanges flags to PageActivity, so that it doesn't get
destroyed and recreated upon screen rotation. This is infinitely better
than our current arrangement, since we'll no longer need to reload all the
content into the WebView, and the WebView itself adjusts its contents to
screen rotation automatically, preserving scroll position.

Also, simplified the layout of the page title and description in the lead
image container. No longer doing the animation and the fade-in of the
description (it's a bit passé).

Bug: T108951
Change-Id: I0c9d0ec21d0ec327981e9d4e02b9c65cd3c3a871