Fix page title position when Wikidata description is empty.
The final piece of the puzzle in correctly laying out the page title and
the Wikidata description below it. This patch fixes the case when the
description is empty, which was causing the title to be totally flush with
the bottom of the lead image.

This also hides the page title until the description is fully laid out,
since the title was sometimes 'jumping' when the page is first loaded.

Change-Id: Ibd3991a8f3615670a697cdc09fdb2442936eddec