Whoops: bump DB version number!
One of our recent patches introduced a new column to one of our
persistence helpers (PageImagePersistenceHelper), which necessitates a
bump of our db version from 6 to 7. In fact, the updated persistence
helper explicitly asks for database version 7, but the actual database
version number wasn't updated.

This is causing the undesired side-effect of images not being saved for
any page (and thumbnails not showing up in History, Saved Pages, etc).

Change-Id: Ie69dee237a86752fe5fd924ce3a253f3b6b2deeb