Hygiene: Add page load performance test
We can't improve what we don't measure.
I'm adding some performance tests to see how long loading a few pages
on testwiki takes. It's good to have those tests added before we significantly
change the page load code further.
For serious testing increase the NUM_RUNS to a higher value.
Also to compare apples to apples it's imperative to test on the same devices,
ideally with the same network condition.
(Ok, the latter is not really exactly controlled. But as long as we're talking
about a stable broadband connection it should be good enough.
The alternative would be mocking responses from a Mock MediaWiki server
and injecting code to use that instead of the real servers.)

Introduced a callback for page load completion which is currently only
used for this performance test but could be used for other things in the future.

Renamed another test to make it clearer that it's testing SectionFetchTask directly.

Bug: T109246
Change-Id: Ie5f8494da952cfdaa753a5e4e7a9f826599a1409