Share only the page URL from the overflow menu.
Prior to this patch, the behavior when selecting the Share option from the
overflow menu was to launch the Share-a-fact dialog with the first
paragraph of text. But in fact, all that users really want is to share the
URL of the page. At the time of this writing, this is one of the most
requested features to be brought back.

Bug: T108181
Change-Id: I371bff4bf5c2b3aeec6dfff708b4de5cba3306cf