Remove RTL icons for social networks, mark SVGs as noflip.
Currently we flip the icons for Twitter, Facebook, and Messaging. As noted
in T109209, this does not make sense. This patch deletes the RTL icons for
those social networks, so that the same, standard icon is used in both LTR
and RTL, and marks the SVGs that generate those icons as noflip, so that RTL
versions are never generated.

Bug: T109209
Change-Id: I6426c69231868212ae7876bd3eea8e51419918cc