Localisation updates from https://translatewiki.net.
A bug in HHVM, T109402, causes most of the line endings to disappear in
translations. After exporting from TWN, the following commands were
executed to printy print the XML:

  find -name strings.xml|
  grep -Ev values/strings.xml|
  xargs -rd\\n -I{} sh -c 'xmllint --format "$1" -o "$1"' -- '{}'

TranslationTests.testAllTranslations() failed because a format specifier
was added to one string. This failure is a false positive as it is
always safe to add a format specifier, but not usually safe to remove
one without updating all translations since the argument would not be
supplied in the Java code. As a workaround, the tests pass after
running:

  find -name strings.xml|
  grep -Ev "values/|-qq/"|
  xargs -rd\\n  sed -ri 's%<string name="toast_saved_page"[^>]*>[^<]*</string>%%'

Change-Id: Ic46d2be21ff97730f06d9d9eb36125482bf19e0c