Fix possible crash involving Zero banner.
When showing the Wikipedia Zero banner from PageActivity, we were
anchoring the SnackBar to a View that's only available when a PageFragment
is active. Therefore, when a different fragment is active (e.g. History),
the view doesn't exist, and thus causes a crash.

This updates the code to use a more generic View from the current Activity
for displaying the banner.

Bug: T110241
Change-Id: I5852b59dd706e428381d8e7b5d39cf414236fe18